Starting to generate King of the Mountain league

+++ backend/kendo-tournament-core/src/test/java/com/softwaremagico/kt/core/tests/ScoreTest.java
@@ -194,7 +194,7 @@ public class ScoreTest extends AbstractTestNGSpringContextTests {
             }
         }
 
-        Assert.assertEquals((int) TEAMS, teamProvider.count(tournament));
+        Assert.assertEquals(TEAMS, teamProvider.count(tournament));
     }
 
     @Test(dependsOnMethods = {"addTeams"})
